# docs/openapi.yaml
openapi: 3.0.3
info:
  title: FUTO-Ride API
  version: 1.0.0
  description: |
    Backend for FUTO-Ride — campus keke pooling, bus tracking, payments (Partna
    NGN onramp → USDC treasury), safety incidents and Telegram alerts.

    Generated from `apps/api/src/routes` + `apps/api/src/schemas`. Zod schemas are the
    source of truth; this document mirrors them for the frontend.

    ## Conventions
    - **Envelope.** Every response is `{ ok: true, data: <T> }` on success or
      `{ ok: false, error: string }` on failure. `data` shapes are the `*Data` schemas below.
    - **Money.** Every amount is **integer kobo** (1 naira = 100 kobo). The frontend renders
      kobo→naira for display and sends kobo back. The backend converts to naira only internally
      at the Partna payment edge.
    - **Auth.** Most endpoints require a Firebase ID token: `Authorization: Bearer <idToken>`.
      Webhooks use their own header/signature auth (not a Firebase token). A few read-only
      endpoints are public.
    - **Validation errors** return **400** `{ ok:false, error }` with `error` shaped as
      `"<field>: <message>"` (first Zod issue), e.g. `"toStop: from and to must differ"`.
    - **Auth failures** return **401** with the exact string `"Missing authentication token"`
      (no bearer token) or `"Invalid or expired token"` (token present but rejected). The
      Partna and Telegram webhooks instead return the literal `"Unauthorized"` on a bad
      signature/secret.
...
servers:
...
security:
  - firebaseAuth: []
paths:
...
  /rides/history:
    get:
      tags: [Rides]
      summary: My past rides (rider)
      description: |
        The caller's **terminal** rides (`completed`, `cancelled`, `expired`), newest first.
        Cursor-paginated on `createdAt` (pass the previous page's `nextCursor` back as `cursor`).
        The rider's in-progress ride is not included here (track it from the booking response).
      parameters:
        - $ref: "#/components/parameters/HistoryLimit"
        - $ref: "#/components/parameters/HistoryCursor"
      responses:
        "200":
          description: Past rides page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiderRideHistoryResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /rides:
    post:
      tags: [Rides]
      summary: Book a keke ride
      description: |
        Books a pooled keke seat (§20). Matches an online keke on the same lane (joining an
        existing pool or opening a new one). One active ride per rider. If no keke is
        available the ride is recorded as **stranded** and returned with `stranded:true`,
        `driverId:null` (HTTP 200, not 409) so the rider can track/cancel it.
        **Rate-limited** (30/min per user).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookRide"
      responses:
        "200":
          description: Ride booked (matched or stranded)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookRideResponse"
        "400":
          description: Unknown stop, or from == to
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              examples:
                unknownStop:
                  summary: fromStop/toStop is not a known campus stop
                  value: { ok: false, error: "Unknown stop" }
                sameStop:
                  summary: fromStop equals toStop (Zod refine)
                  value: { ok: false, error: "toStop: from and to must differ" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: Already have an active ride, or the keke just filled/left
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              examples:
                activeRide:
                  summary: Rider already has an in-progress ride
                  value: { ok: false, error: "You already have an active ride" }
                kekeOffline:
                  summary: Matched keke went offline mid-transaction
                  value: { ok: false, error: "Keke just went offline, try again" }
                kekeFilled:
                  summary: Matched keke filled up / left the lane mid-transaction
                  value: { ok: false, error: "Keke just filled up, try again" }
        "429":
          $ref: "#/components/responses/RateLimited"
  /rides/{id}/cancel:
    post:
      tags: [Rides]
      summary: Cancel a ride
      description: |
        Cancels a ride (§20.4). Callable by the rider OR the assigned driver.
        - **Rider cancels:** tiered refund policy — full refund pre-arrival, partial
          (minus a cancellation fee) once the driver is `arriving`, none mid-trip.
        - **Driver cancels:** the rider is auto re-matched to another keke on the same lane,
          or stranded with `refundPending` if none is available.

        No request body required.
      parameters:
        - $ref: "#/components/parameters/RideId"
      responses:
        "200":
          description: Ride cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CancelResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Not your ride
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { ok: false, error: "Not your ride" }
        "404":
          description: Ride not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { ok: false, error: "Ride not found" }
        "409":
          description: Ride already closed (completed/cancelled/expired)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { ok: false, error: "Ride already closed" }
  /rides/{id}/status:
    post:
      tags: [Rides]
      summary: Advance ride status (driver)
      description: |
        Driver advances the trip through mid-ride states: `assigned → arriving → started`
        (§20.11). One tap advances every PAID pooled rider at the same stage; unpaid peers
        stay put. Requires the ride to be PAID (402 otherwise).
      parameters:
        - $ref: "#/components/parameters/RideId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRideStatus"
      responses:
        "200":
          description: Status advanced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "402":
          description: Rider hasn't paid yet
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { ok: false, error: "Rider hasn't paid yet" }
        "403":
          description: Not your ride
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { ok: false, error: "Not your ride" }
        "404":
          description: Ride not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { ok: false, error: "Ride not found" }
        "409":
          description: Ride closed, or illegal transition
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              examples:
                closed:
                  summary: Ride already completed/cancelled/expired
                  value: { ok: false, error: "Ride already closed" }
                illegalTransition:
                  summary: Not the legal next state (message interpolates the actual states)
                  value: { ok: false, error: "Cannot move from assigned to started" }
  /rides/{id}/complete:
    post:
      tags: [Rides]
      summary: Complete a ride (rider)
      description: |
        The rider completes the trip by presenting the driver's QR token OR the typed PIN —
        exactly one (§20.3). Requires the ride to be `started` and PAID. **Rate-limited**
        (30/min per user).
      parameters:
        - $ref: "#/components/parameters/RideId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteRide"
      responses:
        "200":
          description: Ride completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompleteResponse"
        "400":
          description: Invalid QR code or PIN (or not exactly one supplied)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              examples:
                invalidCode:
                  summary: Supplied QR token / PIN did not match
                  value: { ok: false, error: "Invalid QR code or PIN" }
                notExactlyOne:
                  summary: Neither or both of qrToken/pin supplied (Zod refine)
                  value: { ok: false, error: "qrToken: provide exactly one of qrToken or pin" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "402":
          description: Payment not confirmed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { ok: false, error: "Payment not confirmed" }
        "403":
          description: Only the rider can complete
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { ok: false, error: "Only the rider can complete" }
        "404":
          description: Ride not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { ok: false, error: "Ride not found" }
        "409":
          description: Ride already completed, or not started yet
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              examples:
                alreadyCompleted:
                  summary: Ride is already completed
                  value: { ok: false, error: "Ride already completed" }
                notStarted:
                  summary: Ride has not reached `started`
                  value: { ok: false, error: "Ride not started yet" }
        "429":
          $ref: "#/components/responses/RateLimited"
  /rides/{id}/qr:
    get:
      tags: [Rides]
      summary: Get ride QR token + PIN (driver)
      description: Returns the completion QR token and PIN for the driver to show the rider. Driver-only.
      parameters:
        - $ref: "#/components/parameters/RideId"
      responses:
        "200":
          description: QR + PIN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QrResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Not your ride
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { ok: false, error: "Not your ride" }
        "404":
          description: Ride not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { ok: false, error: "Ride not found" }
  /rides/{id}/rate:
    post:
      tags: [Rides]
      summary: Rate a completed ride (rider)
      description: One rating per ride (rating doc id = ride id). Requires a completed ride. Rider-only.
      parameters:
        - $ref: "#/components/parameters/RideId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RateRide"
      responses:
        "200":
          description: Rating recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Only the rider can rate
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { ok: false, error: "Only the rider can rate" }
        "404":
          description: Ride not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { ok: false, error: "Ride not found" }
        "409":
          description: Ride not completed, or already rated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              examples:
                notCompleted:
                  summary: Ride is not yet completed
                  value: { ok: false, error: "Rate only after the ride is completed" }
                alreadyRated:
                  summary: A rating already exists for this ride
                  value: { ok: false, error: "Ride already rated" }
...
components:
  securitySchemes:
    firebaseAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Firebase ID token — `Authorization: Bearer <idToken>`."
    telegramSecret:
      type: apiKey
      in: header
      name: X-Telegram-Bot-Api-Secret-Token
      description: Shared secret configured on the Telegram bot webhook.
  parameters:
    RideId:
      name: id
      in: path
      required: true
      description: Ride id
      schema: { type: string, minLength: 1 }
...
    HistoryLimit:
      name: limit
      in: query
      required: false
      description: Page size (1–50).
      schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
    HistoryCursor:
      name: cursor
      in: query
      required: false
      description: "`createdAt` (epoch ms) of the last item from the previous page — results page after it."
      schema: { type: integer, minimum: 1 }
  responses:
    Unauthorized:
      description: Missing or invalid Firebase token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          examples:
            missing:
              summary: No bearer token supplied
              value: { ok: false, error: "Missing authentication token" }
            invalid:
              summary: Token present but invalid/expired
              value: { ok: false, error: "Invalid or expired token" }
...
  schemas:
...
    RiderRideHistoryResponse:
      type: object
      required: [ok, data]
      properties:
        ok: { type: boolean, enum: [true] }
        data:
          type: object
          required: [rides, nextCursor]
          properties:
            rides:
              type: array
              items:
                allOf:
                  - $ref: "#/components/schemas/RideHistoryItem"
                  - type: object
                    properties:
                      driverId:
                        type: string
                        nullable: true
                        description: The keke driver, or `null` for a ride that never matched (expired/cancelled while stranded)
            nextCursor:
              type: integer
              nullable: true
              description: Pass back as `cursor` for the next page; `null` when there are no more.
...
    BookRide:
      type: object
      required: [fromStop, toStop]
      properties:
        fromStop: { type: string, minLength: 1, example: seet }
        toStop: { type: string, minLength: 1, example: town }
        payMethod: { type: string, enum: [naira], default: naira }
        priorityFee:
          type: integer
          minimum: 0
          maximum: 500000
          description: Optional surge priority fee in kobo (capped). Honored only when surge is on.
        seats:
          type: integer
          minimum: 1
          maximum: 4
          default: 1
          description: Seats to book (4 = charter the whole keke)
      description: "`fromStop` and `toStop` must differ."
...